requirements.txt
# Core dependencies
streamlit>=1.28.0
pandas>=2.0.0
numpy>=1.24.0
plotly>=5.17.0
python-dotenv>=1.0.0

# LLM API clients
openai>=1.0.0
anthropic>=0.7.0
llamaapi>=0.1.0
httpx>=0.25.0
aiohttp>=3.9.0

# Data processing
openpyxl>=3.1.0
xlsxwriter>=3.1.0
scipy>=1.11.0
scikit-learn>=1.3.0
pydantic>=2.0.0

# Async and utilities
nest-asyncio>=1.5.0
asyncio-throttle>=1.0.2
tenacity>=8.2.0
tiktoken>=0.5.0
tqdm>=4.65.0

# Visualization
seaborn>=0.12.0
matplotlib>=3.7.0
wordcloud>=1.9.0

# Statistical analysis
statsmodels>=0.14.0

# Development
notebook>=7.0.0